@@ -22,6 +22,9 @@ use arrow_schema::*;
 use datafusion_common::{DFSchema, Result, TableReference};
 use datafusion_expr::test::function_stub::{count_udaf, max_udaf, min_udaf, sum_udaf};
 use datafusion_expr::{col, lit, table_scan, wildcard, LogicalPlanBuilder};
+use datafusion_functions::unicode;
+use datafusion_functions_aggregate::grouping::grouping_udaf;
+use datafusion_functions_window::rank::rank_udwf;
 use datafusion_sql::planner::{ContextProvider, PlannerContext, SqlToRel};
 use datafusion_sql::unparser::dialect::{
     DefaultDialect as UnparserDefaultDialect, Dialect as UnparserDialect,
@@ -71,7 +74,7 @@ fn roundtrip_expr() {
 
         let ast = expr_to_sql(&expr)?;
 
-        Ok(format!("{}", ast))
+        Ok(ast.to_string())
     };
 
     for (table, query, expected) in tests {
@@ -139,6 +142,13 @@ fn roundtrip_statement() -> Result<()> {
             SELECT j2_string as string FROM j2
             ORDER BY string DESC
             LIMIT 10"#,
+            r#"SELECT col1, id FROM (
+                SELECT j1_string AS col1, j1_id AS id FROM j1
+                UNION ALL
+                SELECT j2_string AS col1, j2_id AS id FROM j2
+                UNION ALL
+                SELECT j3_string AS col1, j3_id AS id FROM j3
+            ) AS subquery GROUP BY col1, id ORDER BY col1 ASC, id ASC"#,
             "SELECT id, count(*) over (PARTITION BY first_name ROWS BETWEEN UNBOUNDED PRECEDING AND UNBOUNDED FOLLOWING),
             last_name, sum(id) over (PARTITION BY first_name ROWS BETWEEN UNBOUNDED PRECEDING AND UNBOUNDED FOLLOWING),
             first_name from person",
@@ -192,7 +202,7 @@ fn roundtrip_statement() -> Result<()> {
 
         let roundtrip_statement = plan_to_sql(&plan)?;
 
-        let actual = format!("{}", &roundtrip_statement);
+        let actual = &roundtrip_statement.to_string();
         println!("roundtrip sql: {actual}");
         println!("plan {}", plan.display_indent());
 
@@ -224,7 +234,7 @@ fn roundtrip_crossjoin() -> Result<()> {
 
     let roundtrip_statement = plan_to_sql(&plan)?;
 
-    let actual = format!("{}", &roundtrip_statement);
+    let actual = &roundtrip_statement.to_string();
     println!("roundtrip sql: {actual}");
     println!("plan {}", plan.display_indent());
 
@@ -237,7 +247,7 @@ fn roundtrip_crossjoin() -> Result<()> {
         \n    TableScan: j1\
         \n    TableScan: j2";
 
-    assert_eq!(format!("{plan_roundtrip}"), expected);
+    assert_eq!(plan_roundtrip.to_string(), expected);
 
     Ok(())
 }
@@ -478,7 +488,7 @@ fn roundtrip_statement_with_dialect() -> Result<()> {
         let unparser = Unparser::new(&*query.unparser_dialect);
         let roundtrip_statement = unparser.plan_to_sql(&plan)?;
 
-        let actual = format!("{}", &roundtrip_statement);
+        let actual = &roundtrip_statement.to_string();
         println!("roundtrip sql: {actual}");
         println!("plan {}", plan.display_indent());
 
@@ -508,7 +518,7 @@ Projection: unnest_placeholder(unnest_table.struct_col).field1, unnest_placehold
     Projection: unnest_table.struct_col AS unnest_placeholder(unnest_table.struct_col), unnest_table.array_col AS unnest_placeholder(unnest_table.array_col), unnest_table.struct_col, unnest_table.array_col
       TableScan: unnest_table"#.trim_start();
 
-    assert_eq!(format!("{plan}"), expected);
+    assert_eq!(plan.to_string(), expected);
 
     Ok(())
 }
@@ -528,7 +538,7 @@ fn test_table_references_in_plan_to_sql() {
             .unwrap();
         let sql = plan_to_sql(&plan).unwrap();
 
-        assert_eq!(format!("{}", sql), expected_sql)
+        assert_eq!(sql.to_string(), expected_sql)
     }
 
     test(
@@ -558,7 +568,7 @@ fn test_table_scan_with_no_projection_in_plan_to_sql() {
             .build()
             .unwrap();
         let sql = plan_to_sql(&plan).unwrap();
-        assert_eq!(format!("{}", sql), expected_sql)
+        assert_eq!(sql.to_string(), expected_sql)
     }
 
     test(
@@ -657,7 +667,12 @@ where
         .unwrap();
 
     let context = MockContextProvider {
-        state: MockSessionState::default(),
+        state: MockSessionState::default()
+            .with_aggregate_function(sum_udaf())
+            .with_aggregate_function(max_udaf())
+            .with_aggregate_function(grouping_udaf())
+            .with_window_function(rank_udwf())
+            .with_scalar_function(Arc::new(unicode::substr().as_ref().clone())),
     };
     let sql_to_rel = SqlToRel::new(&context);
     let plan = sql_to_rel.sql_statement_to_plan(statement).unwrap();
@@ -667,27 +682,103 @@ where
 }
 
 #[test]
-fn test_table_scan_pushdown() -> Result<()> {
+fn test_table_scan_alias() -> Result<()> {
     let schema = Schema::new(vec![
         Field::new("id", DataType::Utf8, false),
         Field::new("age", DataType::Utf8, false),
     ]);
 
+    let plan = table_scan(Some("t1"), &schema, None)?
+        .project(vec![col("id")])?
+        .alias("a")?
+        .build()?;
+    let sql = plan_to_sql(&plan)?;
+    assert_eq!(sql.to_string(), "SELECT * FROM (SELECT t1.id FROM t1) AS a");
+
+    let plan = table_scan(Some("t1"), &schema, None)?
+        .project(vec![col("id")])?
+        .alias("a")?
+        .build()?;
+
+    let sql = plan_to_sql(&plan)?;
+    assert_eq!(sql.to_string(), "SELECT * FROM (SELECT t1.id FROM t1) AS a");
+
+    let plan = table_scan(Some("t1"), &schema, None)?
+        .filter(col("id").gt(lit(5)))?
+        .project(vec![col("id")])?
+        .alias("a")?
+        .build()?;
+    let sql = plan_to_sql(&plan)?;
+    assert_eq!(
+        sql.to_string(),
+        "SELECT * FROM (SELECT t1.id FROM t1 WHERE (t1.id > 5)) AS a"
+    );
+
+    let table_scan_with_two_filter = table_scan_with_filters(
+        Some("t1"),
+        &schema,
+        None,
+        vec![col("id").gt(lit(1)), col("age").lt(lit(2))],
+    )?
+    .project(vec![col("id")])?
+    .alias("a")?
+    .build()?;
+    let table_scan_with_two_filter = plan_to_sql(&table_scan_with_two_filter)?;
+    assert_eq!(
+        table_scan_with_two_filter.to_string(),
+        "SELECT * FROM (SELECT t1.id FROM t1 WHERE ((t1.id > 1) AND (t1.age < 2))) AS a"
+    );
+
+    let table_scan_with_fetch =
+        table_scan_with_filter_and_fetch(Some("t1"), &schema, None, vec![], Some(10))?
+            .project(vec![col("id")])?
+            .alias("a")?
+            .build()?;
+    let table_scan_with_fetch = plan_to_sql(&table_scan_with_fetch)?;
+    assert_eq!(
+        table_scan_with_fetch.to_string(),
+        "SELECT * FROM (SELECT t1.id FROM (SELECT * FROM t1 LIMIT 10)) AS a"
+    );
+
+    let table_scan_with_pushdown_all = table_scan_with_filter_and_fetch(
+        Some("t1"),
+        &schema,
+        Some(vec![0, 1]),
+        vec![col("id").gt(lit(1))],
+        Some(10),
+    )?
+    .project(vec![col("id")])?
+    .alias("a")?
+    .build()?;
+    let table_scan_with_pushdown_all = plan_to_sql(&table_scan_with_pushdown_all)?;
+    assert_eq!(
+        table_scan_with_pushdown_all.to_string(),
+        "SELECT * FROM (SELECT t1.id FROM (SELECT t1.id, t1.age FROM t1 WHERE (t1.id > 1) LIMIT 10)) AS a"
+    );
+    Ok(())
+}
+
+#[test]
+fn test_table_scan_pushdown() -> Result<()> {
+    let schema = Schema::new(vec![
+        Field::new("id", DataType::Utf8, false),
+        Field::new("age", DataType::Utf8, false),
+    ]);
     let scan_with_projection =
         table_scan(Some("t1"), &schema, Some(vec![0, 1]))?.build()?;
     let scan_with_projection = plan_to_sql(&scan_with_projection)?;
     assert_eq!(
-        format!("{}", scan_with_projection),
+        scan_with_projection.to_string(),
         "SELECT t1.id, t1.age FROM t1"
     );
 
     let scan_with_projection = table_scan(Some("t1"), &schema, Some(vec![1]))?.build()?;
     let scan_with_projection = plan_to_sql(&scan_with_projection)?;
-    assert_eq!(format!("{}", scan_with_projection), "SELECT t1.age FROM t1");
+    assert_eq!(scan_with_projection.to_string(), "SELECT t1.age FROM t1");
 
     let scan_with_no_projection = table_scan(Some("t1"), &schema, None)?.build()?;
     let scan_with_no_projection = plan_to_sql(&scan_with_no_projection)?;
-    assert_eq!(format!("{}", scan_with_no_projection), "SELECT * FROM t1");
+    assert_eq!(scan_with_no_projection.to_string(), "SELECT * FROM t1");
 
     let table_scan_with_projection_alias =
         table_scan(Some("t1"), &schema, Some(vec![0, 1]))?
@@ -696,7 +787,7 @@ fn test_table_scan_pushdown() -> Result<()> {
     let table_scan_with_projection_alias =
         plan_to_sql(&table_scan_with_projection_alias)?;
     assert_eq!(
-        format!("{}", table_scan_with_projection_alias),
+        table_scan_with_projection_alias.to_string(),
         "SELECT ta.id, ta.age FROM t1 AS ta"
     );
 
@@ -707,7 +798,7 @@ fn test_table_scan_pushdown() -> Result<()> {
     let table_scan_with_projection_alias =
         plan_to_sql(&table_scan_with_projection_alias)?;
     assert_eq!(
-        format!("{}", table_scan_with_projection_alias),
+        table_scan_with_projection_alias.to_string(),
         "SELECT ta.age FROM t1 AS ta"
     );
 
@@ -717,7 +808,7 @@ fn test_table_scan_pushdown() -> Result<()> {
     let table_scan_with_no_projection_alias =
         plan_to_sql(&table_scan_with_no_projection_alias)?;
     assert_eq!(
-        format!("{}", table_scan_with_no_projection_alias),
+        table_scan_with_no_projection_alias.to_string(),
         "SELECT * FROM t1 AS ta"
     );
 
@@ -729,7 +820,7 @@ fn test_table_scan_pushdown() -> Result<()> {
     let query_from_table_scan_with_projection =
         plan_to_sql(&query_from_table_scan_with_projection)?;
     assert_eq!(
-        format!("{}", query_from_table_scan_with_projection),
+        query_from_table_scan_with_projection.to_string(),
         "SELECT * FROM (SELECT t1.id, t1.age FROM t1)"
     );
 
@@ -742,7 +833,7 @@ fn test_table_scan_pushdown() -> Result<()> {
     .build()?;
     let table_scan_with_filter = plan_to_sql(&table_scan_with_filter)?;
     assert_eq!(
-        format!("{}", table_scan_with_filter),
+        table_scan_with_filter.to_string(),
         "SELECT * FROM t1 WHERE (t1.id > t1.age)"
     );
 
@@ -755,7 +846,7 @@ fn test_table_scan_pushdown() -> Result<()> {
     .build()?;
     let table_scan_with_two_filter = plan_to_sql(&table_scan_with_two_filter)?;
     assert_eq!(
-        format!("{}", table_scan_with_two_filter),
+        table_scan_with_two_filter.to_string(),
         "SELECT * FROM t1 WHERE ((t1.id > 1) AND (t1.age < 2))"
     );
 
@@ -769,7 +860,7 @@ fn test_table_scan_pushdown() -> Result<()> {
     .build()?;
     let table_scan_with_filter_alias = plan_to_sql(&table_scan_with_filter_alias)?;
     assert_eq!(
-        format!("{}", table_scan_with_filter_alias),
+        table_scan_with_filter_alias.to_string(),
         "SELECT * FROM t1 AS ta WHERE (ta.id > ta.age)"
     );
 
@@ -783,7 +874,7 @@ fn test_table_scan_pushdown() -> Result<()> {
     let table_scan_with_projection_and_filter =
         plan_to_sql(&table_scan_with_projection_and_filter)?;
     assert_eq!(
-        format!("{}", table_scan_with_projection_and_filter),
+        table_scan_with_projection_and_filter.to_string(),
         "SELECT t1.id, t1.age FROM t1 WHERE (t1.id > t1.age)"
     );
 
@@ -797,7 +888,7 @@ fn test_table_scan_pushdown() -> Result<()> {
     let table_scan_with_projection_and_filter =
         plan_to_sql(&table_scan_with_projection_and_filter)?;
     assert_eq!(
-        format!("{}", table_scan_with_projection_and_filter),
+        table_scan_with_projection_and_filter.to_string(),
         "SELECT t1.age FROM t1 WHERE (t1.id > t1.age)"
     );
 
@@ -806,7 +897,7 @@ fn test_table_scan_pushdown() -> Result<()> {
             .build()?;
     let table_scan_with_inline_fetch = plan_to_sql(&table_scan_with_inline_fetch)?;
     assert_eq!(
-        format!("{}", table_scan_with_inline_fetch),
+        table_scan_with_inline_fetch.to_string(),
         "SELECT * FROM t1 LIMIT 10"
     );
 
@@ -821,7 +912,7 @@ fn test_table_scan_pushdown() -> Result<()> {
     let table_scan_with_projection_and_inline_fetch =
         plan_to_sql(&table_scan_with_projection_and_inline_fetch)?;
     assert_eq!(
-        format!("{}", table_scan_with_projection_and_inline_fetch),
+        table_scan_with_projection_and_inline_fetch.to_string(),
         "SELECT t1.id, t1.age FROM t1 LIMIT 10"
     );
 
@@ -835,7 +926,7 @@ fn test_table_scan_pushdown() -> Result<()> {
     .build()?;
     let table_scan_with_all = plan_to_sql(&table_scan_with_all)?;
     assert_eq!(
-        format!("{}", table_scan_with_all),
+        table_scan_with_all.to_string(),
         "SELECT t1.id, t1.age FROM t1 WHERE (t1.id > t1.age) LIMIT 10"
     );
     Ok(())
@@ -892,4 +983,50 @@ fn test_with_offset0() {
 #[test]
 fn test_with_offset95() {
     sql_round_trip(MySqlDialect {}, "select 1 offset 95", "SELECT 1 OFFSET 95");
+}
+
+#[test]
+fn test_order_by_to_sql() {
+    // order by aggregation function
+    sql_round_trip(
+        GenericDialect {},
+        r#"SELECT id, first_name, SUM(id) FROM person GROUP BY id, first_name ORDER BY SUM(id) ASC, first_name DESC, id, first_name LIMIT 10"#,
+        r#"SELECT person.id, person.first_name, sum(person.id) FROM person GROUP BY person.id, person.first_name ORDER BY sum(person.id) ASC NULLS LAST, person.first_name DESC NULLS FIRST, person.id ASC NULLS LAST, person.first_name ASC NULLS LAST LIMIT 10"#,
+    );
+
+    // order by aggregation function alias
+    sql_round_trip(
+        GenericDialect {},
+        r#"SELECT id, first_name, SUM(id) as total_sum FROM person GROUP BY id, first_name ORDER BY total_sum ASC, first_name DESC, id, first_name LIMIT 10"#,
+        r#"SELECT person.id, person.first_name, sum(person.id) AS total_sum FROM person GROUP BY person.id, person.first_name ORDER BY total_sum ASC NULLS LAST, person.first_name DESC NULLS FIRST, person.id ASC NULLS LAST, person.first_name ASC NULLS LAST LIMIT 10"#,
+    );
+
+    // order by scalar function from projection
+    sql_round_trip(
+        GenericDialect {},
+        r#"SELECT id, first_name, substr(first_name,0,5) FROM person ORDER BY id, substr(first_name,0,5)"#,
+        r#"SELECT person.id, person.first_name, substr(person.first_name, 0, 5) FROM person ORDER BY person.id ASC NULLS LAST, substr(person.first_name, 0, 5) ASC NULLS LAST"#,
+    );
+}
+
+#[test]
+fn test_aggregation_to_sql() {
+    sql_round_trip(
+        GenericDialect {},
+        r#"SELECT id, first_name,
+        SUM(id) AS total_sum,
+        SUM(id) OVER (PARTITION BY first_name ROWS BETWEEN 5 PRECEDING AND 2 FOLLOWING) AS moving_sum,
+        MAX(SUM(id)) OVER (PARTITION BY first_name ROWS BETWEEN UNBOUNDED PRECEDING AND UNBOUNDED FOLLOWING) AS max_total,
+        rank() OVER (PARTITION BY grouping(id) + grouping(age), CASE WHEN grouping(age) = 0 THEN id END ORDER BY sum(id) DESC) AS rank_within_parent_1,
+        rank() OVER (PARTITION BY grouping(age) + grouping(id), CASE WHEN (CAST(grouping(age) AS BIGINT) = 0) THEN id END ORDER BY sum(id) DESC) AS rank_within_parent_2
+        FROM person
+        GROUP BY id, first_name;"#,
+        r#"SELECT person.id, person.first_name,
+sum(person.id) AS total_sum, sum(person.id) OVER (PARTITION BY person.first_name ROWS BETWEEN '5' PRECEDING AND '2' FOLLOWING) AS moving_sum,
+max(sum(person.id)) OVER (PARTITION BY person.first_name ROWS BETWEEN UNBOUNDED PRECEDING AND UNBOUNDED FOLLOWING) AS max_total,
+rank() OVER (PARTITION BY (grouping(person.id) + grouping(person.age)), CASE WHEN (grouping(person.age) = 0) THEN person.id END ORDER BY sum(person.id) DESC NULLS FIRST RANGE BETWEEN UNBOUNDED PRECEDING AND CURRENT ROW) AS rank_within_parent_1,
+rank() OVER (PARTITION BY (grouping(person.age) + grouping(person.id)), CASE WHEN (CAST(grouping(person.age) AS BIGINT) = 0) THEN person.id END ORDER BY sum(person.id) DESC NULLS FIRST RANGE BETWEEN UNBOUNDED PRECEDING AND CURRENT ROW) AS rank_within_parent_2
+FROM person
+GROUP BY person.id, person.first_name"#.replace("\n", " ").as_str(),
+    );
 }
